Adds dynamic controller type support
If the controller is registered with a function on the `:keechma.controller/type`
key, this function will be called with a return value of the `:keechma.controller/params`
function. Whatever is returned will be used as a controller type. Returned value
must be a keyword derived from the `:keechma/controller` key.