Test controller lifecycle ordering with subapps
When controller A, a mutual dependee of controller B that lives inside a subapp,
and controller C that is a sibling of controller A (and lives in a parent app of
controller's B app) changes it's state, which causes the controller B to proceed
to shutdown, and controller C to startup - controller's B stop method is invoked
after controller's C start method.

This doesn't happen  when controller B and C are siblings.

stop method gets invoked after parent's start method.