Add keechma.next.event namespace
This namespace holds helper functions that allow you to invert the control over event dispatching or broadcasting. Use them when you want to be able to pass events to the controller from the "outside" (by defining them on the controller map), which controller can trigger without knowing anything about the event (or it's nature - dispatch or broadcast).